fix type warnings

+++ src/wallet.cpp
@@ -1522,8 +1522,8 @@ void CWalletTx::GetAmounts(list<COutputEntry>& listReceived,
 
     CAmount volatile sumIn = 0;
     CAmount volatile sumOut = 0;
-    int volatile numVinFromMe = 0;
-    int volatile numVoutToMe = 0;
+    size_t volatile numVinFromMe = 0;
+    size_t volatile numVoutToMe = 0;
     BOOST_FOREACH (const CTxIn& txin, vin) {
         const auto mi = pwallet->mapWallet.find(txin.prevout.hash);
         if (mi != pwallet->mapWallet.end()) {
